Fix removeBooks add navigation
